Fix an off-by-one error with validation records
PR #32 introduces an obscure off-by-one error that leads to a duplicated
validation record because of a wrap-around in `distinct_domain_names`.

I assume this hasn't been spotted until now mainly because the
duplicated record gets silently overwritten when
`validation_allow_overwrite_records` is set to `true` (as it is by
default).

I tried to identify a corner case when this `+ 1` is required, but
couldn't find any so far. Please let me know if you are aware why it may
be needed.